Add dark mode landing and PDFKit fallback
