+++ apps/web/package.json
@@ -27,6 +27,7 @@
     "lucide-react": "^0.446.0",
     "next": "14.0.0",
     "next-themes": "^0.3.0",
+    "pdfkit": "^0.17.2",
     "puppeteer": "^22.0.0",
     "puppeteer-core": "^22.15.0",
     "react": "^18.2.0",
